% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{to.fractional}
\alias{to.fractional}
\title{Convert decimal odds to fractional odds.}
\usage{
to.fractional(decimal)
}
\arguments{
\item{A}{numbeic vector of decimal odds.}
}
\value{
fractional A fractional character vector of fractional odds.
}
\description{
Convert decimal odds to fractional odds.
}
\examples{
to.decimal(1.5)
}

